+++ GameDevProject/Assets/Scripts/Player/MindSwap.cs
@@ -23,7 +23,7 @@ public class MindSwap : AbilityBase
         playerRenderer.sprite = enemyRenderer.sprite;
         enemyRenderer.sprite = temp;
 
-        float tempHealth = player.health;
+        int tempHealth = player.health;
         player.health = enemy.health;
         enemy.health = tempHealth;
